requirements.txt
PyYAML>=6.0.1
pandas>=2.0.0
rich>=13.0.0
click>=8.1.0 